# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Timetable API",
        "description": "API documentation for the Timetable application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/timetables": {
            "get": {
                "tags": [
                    "Timetables"
                ],
                "summary": "Get list of timetables for the authenticated user",
                "description": "Returns list of timetables",
                "operationId": "getTimetablesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Timetable"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Timetables"
                ],
                "summary": "Store new timetable",
                "description": "Returns timetable data",
                "operationId": "storeTimetable",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Timetable"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Timetable"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/timetables/{id}": {
            "get": {
                "tags": [
                    "Timetables"
                ],
                "summary": "Get timetable information",
                "description": "Returns timetable data",
                "operationId": "getTimetableById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Timetable id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Timetable"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Timetables"
                ],
                "summary": "Update existing timetable",
                "description": "Returns updated timetable data",
                "operationId": "updateTimetable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Timetable id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Timetable"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Timetable"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Timetables"
                ],
                "summary": "Delete existing timetable",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteTimetable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Timetable id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/timetable-preferences": {
            "post": {
                "tags": [
                    "Timetable Preferences"
                ],
                "summary": "Store or update user's timetable preferences",
                "description": "Returns the created or updated preference data",
                "operationId": "storeTimetablePreference",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TimetablePreferenceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimetablePreference"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Day": {
                "title": "Day",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Monday"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
            "Lecturer": {
                "title": "Lecturer",
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "123-456-7890"
                    },
                    "department": {
                        "type": "string",
                        "example": "Computer Science"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
            "Subject": {
                "title": "Subject",
                "required": [
                    "name",
                    "code"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Introduction to Programming"
                    },
                    "code": {
                        "type": "string",
                        "example": "CS101"
                    },
                    "description": {
                        "type": "string",
                        "example": "An introductory course on programming fundamentals."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "TimeSlot": {
                "title": "TimeSlot",
                "required": [
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "time",
                        "example": "09:00:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "time",
                        "example": "10:00:00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
            "Timetable": {
                "title": "Timetable",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_by": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "timetable_entries": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TimetableEntry"
                        }
                    }
                },
                "type": "object"
            },
            "TimetablePreferenceRequest": {
                "title": "Timetable Preference Request",
                "required": [
                    "preferences"
                ],
                "properties": {
                    "preferences": {
                        "description": "A key-value object of user preferences for timetable generation.",
                        "type": "object",
                        "example": {
                            "prefer_morning": true,
                            "max_consecutive_classes": 3
                        }
                    }
                },
                "type": "object"
            },
            "TimetableEntry": {
                "title": "Timetable Entry",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "timetable_id": {
                        "type": "integer"
                    },
                    "subject_id": {
                        "type": "integer"
                    },
                    "lecturer_id": {
                        "type": "integer"
                    },
                    "day_id": {
                        "type": "integer"
                    },
                    "time_slot_id": {
                        "type": "integer"
                    },
                    "activity": {
                        "type": "string"
                    },
                    "section": {
                        "type": "string"
                    },
                    "venue": {
                        "type": "string"
                    },
                    "tied_to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "subject": {
                        "$ref": "#/components/schemas/Subject"
                    },
                    "lecturer": {
                        "$ref": "#/components/schemas/Lecturer"
                    },
                    "day": {
                        "$ref": "#/components/schemas/Day"
                    },
                    "time_slot": {
                        "$ref": "#/components/schemas/TimeSlot"
                    }
                },
                "type": "object"
            },
            "TimetablePreference": {
                "title": "Timetable Preference",
                "required": [
                    "preferences"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "user_id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "preferences": {
                        "type": "object",
                        "additionalProperties": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}